Updated chart-loader for manager dashboard
