+++ packages/manager/src/pages/dashboard/components/dashboard-portfolio-chart-section/dashboard-chart/dashboard-portfolio-chart-container.js
@@ -12,6 +12,8 @@ import {
 } from "../../../services/dashboard.service";
 import DashboardChartLoader
   from "shared/components/dashboard/dashboard-chart-loader/dashboard-chart-loader";
+import DashboardChartDescriptionLoader
+  from "shared/components/dashboard/dashboard-chart-loader/dashboard-chart-description-loader";
 
 class DashboardPortfolioChartContainer extends PureComponent {
   componentDidMount() {
@@ -28,8 +30,12 @@ class DashboardPortfolioChartContainer extends PureComponent {
   render() {
     const { assetChart, currency, period } = this.props;
     if (!assetChart)
-      return <DashboardChartLoader />;
-    if (!assetChart) return null;
+      return (
+        <Fragment>
+          <DashboardChartDescriptionLoader />
+          <DashboardChartLoader />
+        </Fragment>
+      );
     return (
       <Fragment>
         <h3 className="dashboard-portfolio-chart-section__heading">
@@ -62,7 +68,6 @@ const mapStateToProps = state => {
   const { assetChart, period } = state.dashboard;
   const { currency } = state.accountSettings;
   const { programs, funds } = state.dashboard;
-  console.log(state.dashboard);
   return {
     assetChart,
     period,
